Added an example config file.
This file can be a txt file or a .json file but it must be in JSON format
(or at least a format that the Python JSON module can understand). Ideally
simply follow the fields outlined in this example .txt file. This ensures
the secure storage of credentials from the internet, should you choose to
publish code such as this. Ideally make sure to store this .txt or .json
file somewhere relatively secure on your PC too.