Committing drivers to install before using the USB to Serial Adapter.
These drivers were not created by me, they are the property of Maxim and
Prolific. This Git repository can serve as a mirror for the
Prolific drivers as their site could not be accessed last I tried. These
drivers are freely available on the internet. The drivers in this
directory include 1-Wire drivers and USB-to-Serial drivers for the
DS9481R-3C7 part.